refactor: move logic in sep files
