+++ lua/plugins/checkpatch.lua
@@ -1,9 +1,14 @@
-require "parser"
+local utils = require "utils"
+local config = require "config"
+local parser = require "parser"
 
 local M = {}
+-- namespace for diagnostics
+M.ns = vim.api.nvim_create_namespace("checkpatch")
+
 local configured = false
 
-local default_config = {
+local default_key_config = {
     mappings = {
         run = {keys = "<leader>cp", desc = "Run checkpatch"},
         next = {keys = ",", desc = "Next checkpatch remark"},
@@ -13,16 +18,19 @@ local default_config = {
 
 function M.setup(opts)
     if configured then return end
-    local ok_tbl, tbl = pcall(require, "vim.tbl")
-    local cfg
+
+	
+
+    local ok_tbl, _ = pcall(require, "vim.tbl")
+    local key_cfg
     if ok_tbl then
-        cfg = vim.tbl_deep_extend("force", default_config, opts or {})
+        key_cfg = vim.tbl_deep_extend("force", default_key_config, opts or {})
     else
-        cfg = default_config
+        key_cfg = default_key_config
     end
 
     local map = vim.keymap.set
-    local km = cfg.mappings or {}
+    local km = key_cfg.mappings or {}
     if km.run and km.run.keys then
         map("n", km.run.keys, ":Checkpatch<CR>",
             {silent = true, desc = km.run.desc or "Run checkpatch"})
@@ -43,132 +51,8 @@ end
 
 local current_index = 0
 
--- namespace for diagnostics
-M.ns = vim.api.nvim_create_namespace("checkpatch")
-
--- remember last used flags across runs
-local function get_last_cfg()
-    local cfg = vim.g.checkpatch_last_cfg
-    if type(cfg) ~= "table" then
-        cfg = {
-            strict = false,
-            codespell = false,
-            log = false,
-            no_tree = false,
-            quiet = false
-        }
-    end
-    return cfg
-end
-
-local function set_last_cfg(cfg)
-    vim.g.checkpatch_last_cfg = cfg
-    return cfg
-end
-
-table.filter = function(array, filterIterator)
-    local result = {}
-
-    for key, value in pairs(array) do
-        if filterIterator(value, key, array) then
-            table.insert(result, value)
-        end
-    end
-
-    return result
-end
-
-function DiagnosticIndicator()
-    local counts = vim.diagnostic.get_count(0)
-
-    if not counts or (counts.remark == 0 and counts.warn == 0) then return "" end
-
-    local parts = {}
-    if counts.error and counts.error > 0 then
-        table.insert(parts, " " .. counts.remark)
-    end
-    if counts.warn and counts.warn > 0 then
-        table.insert(parts, " " .. counts.warn)
-    end
-
-    return table.concat(parts, " ")
-end
-
-local function write_log(data)
-    local log_dir = vim.fn.stdpath("data") .. "/checkpatch-logs"
-    vim.fn.mkdir(log_dir, "p") -- create dir if missing
-
-    local filename = log_dir .. "/log_" .. os.date("%Y-%m-%d_%H-%M") .. ".txt"
-    local file, err = io.open(filename, "w")
-    if not file then
-        print("Error opening log file:", err)
-        return
-    end
-
-    file:write(data)
-    file:close()
-end
-
-local function file_exists(path)
-    local f = io.open(path, "r")
-    if f then f:close() end
-    return f ~= nil
-end
-
-local function install_checkpatch()
-    local script_dir = vim.fn.stdpath("data") .. "/checkpatch-scripts"
-    vim.fn.mkdir(script_dir, "p")
-
-    local checkpatch_file = script_dir .. "/checkpatch.pl"
-    local spelling_file = script_dir .. "/spelling.txt"
-    local const_structs_file = script_dir .. "/const_structs.checkpatch"
-
-    -- Download helper (synchronous)
-    local function curl_get(url, dest)
-        vim.fn.system({"curl", "-sSL", "-o", dest, url})
-        if vim.v.shell_error ~= 0 then
-            print("Failed to download: " .. url)
-            return false
-        end
-        return true
-    end
-
-    -- Ensure main script
-    if not file_exists(checkpatch_file) then
-        if curl_get(
-            "https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/plain/scripts/checkpatch.pl",
-            checkpatch_file) then
-            vim.fn.system({"chmod", "+x", checkpatch_file})
-        end
-    end
-
-    -- Ensure aux files that checkpatch expects when certain flags are used
-    if not file_exists(spelling_file) then
-        curl_get(
-            "https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/plain/scripts/spelling.txt",
-            spelling_file)
-    end
-
-    if not file_exists(const_structs_file) then
-        curl_get(
-            "https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/plain/scripts/const_structs.checkpatch",
-            const_structs_file)
-    end
-
-    return checkpatch_file
-end
-
-local function get_remarks()
-    local bufnr = vim.api.nvim_get_current_buf()
-    local diag = vim.diagnostic.get(bufnr)
-    local filteredDiag = table.filter(diag, function(element, key, index)
-        return element.source == "checkpatch"
-    end)
-    return filteredDiag
-end
-
 function M.next_remark()
-    local remarks = get_remarks()
+    local remarks = utils.get_remarks()
     if #remarks == 0 then
         print("No remarks")
         return
@@ -183,7 +67,7 @@ function M.next_remark()
 end
 
 function M.prev_remark()
-    local remarks = get_remarks()
+    local remarks = utils.get_remarks()
     if #remarks == 0 then
         print("No remarks found")
         return
@@ -200,29 +84,18 @@ function M.prev_remark()
     vim.notify(err.message, log_level)
 end
 
--- main execution function
-local function shell_escape_dquote(str) return (str or ""):gsub('"', '\\"') end
-
-local function get_repo_root(start_dir)
-    local out = vim.fn.systemlist({
-        "git", "-C", start_dir, "rev-parse", "--show-toplevel"
-    })
-    if vim.v.shell_error ~= 0 or not out or #out == 0 then return nil end
-    return out[1]
-end
-
 function M.run(cfg)
     local buf = vim.api.nvim_get_current_buf()
     local file = vim.api.nvim_buf_get_name(buf)
     if file == "" then
         if not cfg.quiet then
-            vim.notify("Buffer is empty. Running in current dir",
+            vim.notify("Buffer is empty. Open some file.",
                        vim.log.levels.INFO)
-            cfg.filem = true
+			return
         end
     end
 
-    local checkpatch_path = install_checkpatch()
+    local checkpatch_path = utils.install_checkpatch()
 
     local opts = "--terse "
     if cfg.strict then opts = opts .. "--strict " end
@@ -233,39 +106,32 @@ function M.run(cfg)
 
     local handle
     if cfg.diff then
-        local file_dir = vim.fn.fnamemodify(file, ":p:h")
-        local repo_root = get_repo_root(file_dir)
-        if not repo_root then
-            if not cfg.quiet then
-                vim.notify(
-                    "checkpatch: not a git repo; falling back to file mode",
-                    vim.log.levels.WARN)
-            end
-            handle = io.popen("perl " .. checkpatch_path .. " " .. opts ..
-                                  "--file " .. file)
-        else
-            local base = cfg.diff_base or "HEAD"
-            local esc_repo = '"' .. shell_escape_dquote(repo_root) .. '"'
-            local esc_file = '"' .. shell_escape_dquote(file) .. '"'
-            local cmd = "git -C " .. esc_repo .. " diff " .. base .. " -- " ..
-                            esc_file .. " | perl " .. checkpatch_path .. " " ..
-                            opts .. "-"
-            handle = io.popen(cmd)
-        end
+		local base_commit = cfg.diff_base or ""
+		local cmd = string.format(
+			"git diff %s | perl %s %s --file %s",
+			base_commit,
+			checkpatch_path,
+			opts or "",
+			file
+		)
+		handle = io.popen(cmd)
     else
         handle = io.popen(
                      "perl " .. checkpatch_path .. " " .. opts .. "--file " ..
                          file)
     end
     -- This function is system dependent and is not available on all platforms. (lua 5.1 ref manual)
-    if not handle then print("Error opening the file") end
+	if not handle then
+		vim.notify("Failed to run checkpatch command", vim.log.levels.ERROR)
+		return
+	end
 
     local result = handle:read("*a")
     handle:close()
 
-    if cfg.log then write_log(result) end
+    if cfg.log then utils.write_log(result) end
 
-    local diagnostics = parse_checkpatch(result)
+    local diagnostics = parser.parse_result(result)
 
     -- Reset and set diagnostics (use default Neovim visuals)
     vim.diagnostic.reset(M.ns, buf)
@@ -277,9 +143,14 @@ function M.run(cfg)
 end
 
 vim.api.nvim_create_user_command("Checkpatch", function(opts)
-    local cfg = get_last_cfg()
-
     local args = opts.fargs
+    local cfg = config.get_last_cfg()
+
+	if !vim.tbl_contains(args, "set") then
+		M.run(cfg)
+		return
+	end
+
     if #args > 0 then
         local overrides = {
             strict = vim.tbl_contains(args, "strict"),
@@ -287,7 +158,6 @@ vim.api.nvim_create_user_command("Checkpatch", function(opts)
             log = vim.tbl_contains(args, "log"),
             no_tree = vim.tbl_contains(args, "no-tree"),
             quiet = vim.tbl_contains(args, "quiet"),
-            filem = vim.tbl_contains(args, "check-all"),
             diff = vim.tbl_contains(args, "diff")
         }
         for _, a in ipairs(args) do
@@ -297,20 +167,28 @@ vim.api.nvim_create_user_command("Checkpatch", function(opts)
             if base then overrides.diff_base = base end
         end
         for k, v in pairs(overrides) do cfg[k] = v end
-        cfg = set_last_cfg(cfg)
+		if (vim.tbl_contains(args, "set")) then
+			config.set_last_cfg(cfg)
+		end
     end
 
     M.run(cfg)
+
 end, {desc = "Highlights the checkpatch msg in buf", nargs = "*"})
 
 -- auto-exec on .c safe
 vim.api.nvim_create_autocmd("BufWritePost", {
     pattern = "*.c",
     callback = function()
-        local cfg = get_last_cfg()
+        local cfg = config.get_last_cfg()
         cfg.quiet = true
+		cfg.diff = true
         M.run(cfg)
     end
 })
 
 return M
+
+
+
+
